Split report validation functions.
