+++ assets/js/util/report-validation.test.js
@@ -19,12 +19,7 @@
 /**
  * Internal dependencies
  */
-import {
-	isValidDateRange,
-	isValidDimensions,
-	isValidMetrics,
-	isValidOrders,
-} from './report-validation';
+import { isValidDateRange, isValidOrders } from './report-validation';
 
 describe( 'Reporting API validation', () => {
 	describe( 'isValidDateRange', () => {
@@ -47,95 +42,6 @@ describe( 'Reporting API validation', () => {
 				startDate: '2020',
 				endDate: '2020-01-01',
 			} ) ).toBeFalsy();
-		} );
-	} );
-
-	describe( 'isValidDimensions', () => {
-		it( 'should return TRUE if a non empty string is passed', () => {
-			expect( isValidDimensions( 'test' ) ).toBeTruthy();
-		} );
-
-		it( 'should return TRUE if a valid object is passed', () => {
-			expect( isValidDimensions( {
-				name: 'test',
-			} ) ).toBeTruthy();
-		} );
-
-		it( 'should return TRUE if a valid array of objects/strings is passed', () => {
-			expect( isValidDimensions( [
-				{
-					name: 'test',
-				},
-				'test2',
-				'test3',
-				{
-					name: 'test4',
-				},
-			] ) ).toBeTruthy();
-		} );
-
-		it( 'should return FALSE if neither string nor array is passed', () => {
-			expect( isValidDimensions( 5.2 ) ).toBeFalsy();
-		} );
-
-		it( 'should return FALSE if not a valid array of objects/strings is passed', () => {
-			expect( isValidDimensions( [
-				{
-					name: 'test',
-				},
-				'test2',
-				5,
-				{
-					name: 'test4',
-				},
-			] ) ).toBeFalsy();
-		} );
-	} );
-
-	describe( 'isValidMetrics', () => {
-		it( 'should return TRUE if a non empty string is passed', () => {
-			expect( isValidMetrics( 'test' ) ).toBeTruthy();
-		} );
-
-		it( 'should return TRUE if a valid object is passed', () => {
-			expect( isValidMetrics( {
-				expression: 'test',
-				alias: 'Test',
-			} ) ).toBeTruthy();
-		} );
-
-		it( 'should return TRUE if a valid array of objects/strings is passed', () => {
-			expect( isValidMetrics( [
-				{
-					expression: 'test',
-					alias: 'Test',
-				},
-				'test2',
-				'test3',
-				{
-					expression: 'test4',
-					alias: 'Test4',
-				},
-			] ) ).toBeTruthy();
-		} );
-
-		it( 'should return FALSE if neither string nor array is passed', () => {
-			expect( isValidMetrics( 5.2 ) ).toBeFalsy();
-		} );
-
-		it( 'should return FALSE if not a valid array of objects/strings is passed', () => {
-			expect( isValidMetrics( [
-				{
-					expression: 'test',
-					alias: 'Test',
-				},
-				'test2',
-				5,
-				{
-					expression: 'test4',
-					alias: 'Test4',
-				},
-			] ) ).toBeFalsy();
 		} );
 	} );
